+++ src/modules/currency/wallet/currency-wallet-reducer.js
@@ -4,24 +4,41 @@ import { buildReducer, filterReducer, memoizeReducer } from 'redux-keto'
 
 import type {
   EdgeCurrencyInfo,
+  EdgeTransaction,
   EdgeWalletInfo
 } from '../../../edge-core-index.js'
 import type { RootAction } from '../../actions.js'
 import type { RootState } from '../../root-reducer.js'
 import { getCurrencyInfo } from '../currency-selectors.js'
 
-export type TxIdHash = {
-  [txidHash: string]: number
+/** Maps from txid hash to file creation date & path. */
+export type TxFileNames = {
+  [txidHash: string]: {
+    creationDate: number,
+    fileName: string
+  }
 }
 
-export type TxFileName = {
-  timestamp: number,
-  fileName: string
-}
+/** Maps from txid hash to file contents (in JSON). */
+export type TxFileJsons = { [txidHash: string]: Object }
+
+/** Maps from txid hash to creation date. */
+export type TxidHashes = { [txidHash: string]: number }
 
 export type SortedTransactions = {
   sortedList: Array<string>,
-  txidHashes: TxIdHash
+  txidHashes: TxidHashes
+}
+
+export type MergedTransaction = {
+  blockHeight: number,
+  date: number,
+  ourReceiveAddresses: Array<string>,
+  signedTx: string,
+  txid: string,
+
+  nativeAmount: { [currencyCode: string]: string },
+  networkFee: { [currencyCode: string]: string }
 }
 
 export interface CurrencyWalletState {
@@ -29,15 +46,15 @@ export interface CurrencyWalletState {
   engineFailure: Error | null;
   fiat: string;
   fiatLoaded: boolean;
-  files: { [txidHash: string]: Object };
-  fileNames: { [txidHash: string]: TxFileName };
+  files: TxFileJsons;
+  fileNames: TxFileNames;
   fileNamesLoaded: boolean;
   sortedTransactions: SortedTransactions;
   name: string | null;
   nameLoaded: boolean;
   walletInfo: EdgeWalletInfo;
   txids: Array<string>;
-  txs: { [txid: string]: Object };
+  txs: { [txid: string]: MergedTransaction };
 }
 
 export interface CurrencyWalletNext {
@@ -66,11 +83,10 @@ const currencyWalletReducer = buildReducer({
     return action.type === 'CURRENCY_WALLET_FIAT_CHANGED' ? true : state
   },
 
-  files (state = {}, action: RootAction) {
+  files (state: TxFileJsons = {}, action: RootAction) {
     switch (action.type) {
       case 'CURRENCY_WALLET_FILE_CHANGED': {
-        const { json, txFileName } = action.payload
-        const { txidHash } = txFileName
+        const { json, txidHash } = action.payload
         const out = { ...state }
         out[txidHash] = json
         return out
@@ -86,8 +102,12 @@ const currencyWalletReducer = buildReducer({
     return state
   },
 
-  sortedTransactions (state = {}, action: RootAction, next: CurrencyWalletNext) {
-    const { txidHashes = {} } = state
+  sortedTransactions (
+    state: SortedTransactions = { txidHashes: {}, sortedList: [] },
+    action: RootAction,
+    next: CurrencyWalletNext
+  ) {
+    const { txidHashes } = state
     switch (action.type) {
       case 'CURRENCY_ENGINE_CHANGED_TXS': {
         return sortTxs(txidHashes, action.payload.txidHashes)
@@ -96,7 +116,7 @@ const currencyWalletReducer = buildReducer({
         const { txFileNames } = action.payload
         const newTxidHashes = {}
         Object.keys(txFileNames).map(txidHash => {
-          newTxidHashes[txidHash] = txFileNames[txidHash].timestamp
+          newTxidHashes[txidHash] = txFileNames[txidHash].creationDate
         })
         return sortTxs(txidHashes, newTxidHashes)
       }
@@ -104,7 +124,7 @@ const currencyWalletReducer = buildReducer({
     return state
   },
 
-  fileNames (state = {}, action: RootAction) {
+  fileNames (state: TxFileNames = {}, action: RootAction) {
     switch (action.type) {
       case 'CURRENCY_WALLET_FILE_NAMES_LOADED': {
         const { txFileNames } = action.payload
@@ -114,10 +134,9 @@ const currencyWalletReducer = buildReducer({
         }
       }
       case 'CURRENCY_WALLET_FILE_CHANGED': {
-        const { txFileName } = action.payload
-        const { txidHash, timestamp, fileName } = txFileName
-        if (!state[txidHash] || timestamp < state[txidHash].timestamp) {
-          state[txidHash] = { timestamp, fileName }
+        const { fileName, creationDate, txidHash } = action.payload
+        if (!state[txidHash] || creationDate < state[txidHash].creationDate) {
+          state[txidHash] = { creationDate, fileName }
         }
         return state
       }
@@ -167,18 +186,20 @@ const currencyWalletReducer = buildReducer({
   }
 })
 
-export function sortTxs (txidHashes: TxIdHash, newHashes: TxIdHash) {
+export function sortTxs (txidHashes: TxidHashes, newHashes: TxidHashes) {
   for (const newTxidHash in newHashes) {
     const newTime = newHashes[newTxidHash]
     if (!txidHashes[newTxidHash] || newTime < txidHashes[newTxidHash]) {
       txidHashes[newTxidHash] = newTime
     }
   }
-  const sortedList = Object.keys(txidHashes).sort((txidHash1, txidHash2) => {
-    if (txidHashes[txidHash1] > txidHashes[txidHash2]) return -1
-    if (txidHashes[txidHash1] < txidHashes[txidHash2]) return 1
-    return 0
-  })
+  const sortedList: Array<string> = Object.keys(txidHashes).sort(
+    (txidHash1, txidHash2) => {
+      if (txidHashes[txidHash1] > txidHashes[txidHash2]) return -1
+      if (txidHashes[txidHash1] < txidHashes[txidHash2]) return 1
+      return 0
+    }
+  )
   return { sortedList, txidHashes }
 }
 
@@ -191,10 +212,25 @@ export default filterReducer(
   }
 )
 
+const defaultTx: MergedTransaction = {
+  blockHeight: 0,
+  date: 0,
+  ourReceiveAddresses: [],
+  signedTx: '',
+  txid: '',
+  nativeAmount: {},
+  networkFee: {},
+  providerFee: {}
+}
+
 /**
  * Merges a new incoming transaction with an existing transaction.
  */
-export function mergeTx (tx: any, defaultCurrency: string, oldTx: any = {}) {
+export function mergeTx (
+  tx: EdgeTransaction,
+  defaultCurrency: string,
+  oldTx: MergedTransaction = defaultTx
+): MergedTransaction {
   const out = {
     blockHeight: tx.blockHeight,
     date: tx.date,
@@ -203,18 +239,14 @@ export function mergeTx (tx: any, defaultCurrency: string, oldTx: any = {}) {
     txid: tx.txid,
 
     nativeAmount: { ...oldTx.nativeAmount },
-    networkFee: { ...oldTx.networkFee },
-    providerFee: { ...oldTx.providerFee }
+    networkFee: { ...oldTx.networkFee }
   }
 
   const currencyCode =
     tx.currencyCode != null ? tx.currencyCode : defaultCurrency
-  out.nativeAmount[currencyCode] =
-    tx.amountSatoshi != null ? tx.amountSatoshi.toString() : tx.nativeAmount
+  out.nativeAmount[currencyCode] = tx.nativeAmount
   out.networkFee[currencyCode] =
     tx.networkFee != null ? tx.networkFee.toString() : '0'
-  out.providerFee[currencyCode] =
-    tx.providerFee != null ? tx.providerFee.toString() : '0'
 
   return out
 }