+++ src/modules/exchange/exchange-api.js
@@ -1,8 +1,8 @@
 // @flow
 
 import { wrapObject } from '../../util/api.js'
-import { getExchangeRate } from '../exchange/exchange-selectors.js'
 import type { ApiInput } from '../root.js'
+import { getExchangeRate } from './exchange-selectors.js'
 
 /**
  * Creates an `ExchangeCache` API object.